feat: add FeederSettingModel and FeederSettingController
- Implemented FeederSettingModel to manage feeder settings with properties like settingId, cloudUrl, loraPort, and type.
- Created FeederSettingController to handle feeder settings, including available serial ports and commented-out methods for updating cloud URL, LoRa port, and type.

+++ lib/app/modules/smart_halter/rule_engine/halter_rule_engine_controller.dart
@@ -117,7 +117,7 @@ class HalterRuleEngineController extends GetxController {
     int? logId,
     double? suhu,
     double? spo,
-    int? bpm,
+    double? bpm,
     double? respirasi,
     int? battery,
     DateTime? time,
